requirements.txt
Django==5.2.6
djangorestframework==3.16.1
python-decouple==3.8
Pillow==11.3.0
celery==5.6.0
redis==7.1.0
psycopg2-binary==2.9.11
reportlab==4.4.7
